requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-classic>=1.0.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pypdf>=5.0.0
rank_bm25>=0.2.2
python-dotenv>=1.0.1
redis>=5.0.0
langchain-anthropic>=0.1.0
langchain-aws>=0.1.0
SQLAlchemy>=2.0.0
langchain>=0.2.0
